ml_service/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pymongo>=3.12.0
pandas>=1.3.0
scipy>=1.7.0
implicit>=0.6.0
scikit-learn>=0.24.0
joblib>=1.0.0
python-multipart>=0.0.5
python-dotenv>=0.19.0